Add option to mark task as canceled in things when they are missing from the JQL fetch
When tickets are deleted or reassigned in Jira, they may stop being returned in the JQL response. These would previously sit in Things forever until acted on manually. This adds an option `CANCEL_MISSING_TICKETS` to the config that will cause any tickets that go missing from subsequent JQL fetches to be marked as `canceled` in Things. If they later reappear in the JQL fetch, they will be un-canceled.